steamerapi/goml-kol-pulse-goml-uptempo/goml-kol-pulse-goml-uptempo/requirements.txt
fastapi==0.110.2
python-dotenv 
pyjwt
boto3
mangum==0.17.0
requests
pandas
llama-index
uvicorn
pydantic
llama-index-llms-bedrock
llama-index-embeddings-bedrock
streamlit
aiofiles
numpy
matplotlib
seaborn
adjustText



